refactor: - removed machine logic from hook - refactored conditional color logic to not force a default color - changed `utilities` folder to `utils` to follow conventions
